// src/app/shared/models/champagne.model.ts
export interface Champagne {
    id: string;
    productName?: string;
    price?: number;
    picture: string;
    winery: string;
    unitPrice: number;
    articleNumber: number;
    vintage?: number;
    color: string;
    packaging?: string;
    wineRegion: string;
    country: string;
    grapeVariety: string;
    nature: string;
    wineStyle: string;
    procedure: string;
}